share/discovery: moving the package closer to shrex
Closes #2206
